move PKAddPaymentPassViewControllerDelegate impl into extension
